backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "evidence-seeker-backend"
version = "1.0.0"
description = "Backend API for Evidence Seeker Platform"
authors = [
    {name = "Evidence Seeker Team", email = "team@evidence-seeker.com"}
]
readme = "README.md"  # backend-local README (added to satisfy build step inside backend context)
requires-python = ">=3.12"
dependencies = [
    # Web framework & ASGI server
    "fastapi==0.116.1",
    "uvicorn[standard]==0.35.0",

    # Database & ORM
    "sqlalchemy==2.0.43",
    "alembic==1.16.5",
    "psycopg2==2.9.10",        # compiled (preferred for production) over psycopg2-binary
    "asyncpg==0.30.0",
    "pgvector==0.4.1",

    # Auth & security
    "fastapi-users[sqlalchemy]==14.0.1",
    "python-jose[cryptography]==3.5.0",
    "passlib[bcrypt]==1.7.4",
    "bcrypt==4.3.0",
    "python-multipart==0.0.20",
    "cryptography==45.0.7",

    # Settings & validation
    "pydantic==2.11.9",
    "pydantic-settings==2.10.1",

    # Logging
    "loguru==0.7.3",

    # Email / templating
    "fastapi-mail==1.4.1",
    "jinja2==3.1.2",

    # AI / embeddings / vector search (CPU only torch)
    "torch==2.4.0",
    "torchvision==0.19.0",
    "evidence-seeker @ git+https://github.com/debatelab/evidence-seeker.git@main",
]

[project.optional-dependencies]
dev = [
    "pytest==8.4.1",
    "pytest-asyncio==1.1.0",
    "pytest-cov==4.1.0",
    "pytest-timeout==2.3.1",
    "black==25.1.0",
    "ruff==0.6.9",  # bring ruff closer to current while not latest edge
    "mypy==1.17.1",
    "pre-commit==3.8.0",
    "httpx==0.28.1",  # used in tests
    "types-passlib>=1.7.7.20250602",
]

[tool.setuptools]
packages = ["app"]

[tool.black]
line-length = 88
target-version = ['py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"tests/*" = ["B011"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true
warn_redundant_casts = true
warn_unused_ignores = true
show_error_codes = true
# Avoid namespace package mode; treat 'app' as a regular package with __init__.py
# (CI had trouble resolving modules when namespace_packages was enabled.)
# namespace_packages = true
## explicit_package_bases can cause package discovery issues in simple monorepos;
## rely on default discovery when running mypy from the backend folder.
## explicit_package_bases = true

[[tool.mypy.overrides]]
module = [
    "fastapi_users.*",
    "loguru",
]
ignore_missing_imports = true

## Tests: relax strict function typing to avoid noise in test code
[[tool.mypy.overrides]]
module = [
    "tests.*",
]
disallow_untyped_defs = false
check_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--strict-markers --strict-config --cov=app --cov-report=term-missing"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]

[dependency-groups]
dev = [
    "pytest-timeout>=2.3.1",
    "types-passlib>=1.7.7.20250602",
]
